Move normalize_phone_number out of models
