+++ funnel/utils.py
@@ -7,10 +7,36 @@ import urllib.parse
 
 from flask import abort
 
+import phonenumbers
 import qrcode
 import qrcode.image.svg
 
+# Unprefixed phone numbers are assumed to be a local number in India (+91) or US (+1).
+# Both IN and US numbers are 10 digits before prefixes. We try IN first as it's the
+# higher priority home region.
+PHONE_LOOKUP_REGIONS = ['IN', 'US']
+
 # --- Utilities ------------------------------------------------------------------------
+
+
+def normalize_phone_number(candidate: str) -> Optional[str]:
+    """Attempt to parse a phone number from a candidate and return in E164 format."""
+    # Assume unprefixed numbers to be a local number in one of the supported common
+    # regions. We start with the higher priority home region and return the _first_
+    # candidate that is likely to be a valid number. This behaviour differentiates it
+    # from similar code in :func:`~funnel.models.utils.getuser`, where the loop exits
+    # with the _last_ valid candidate (as it's coupled with a
+    # :class:`~funnel.models.user.UserPhone` lookup)
+    try:
+        for region in PHONE_LOOKUP_REGIONS:
+            parsed_number = phonenumbers.parse(candidate, region)
+            if phonenumbers.is_valid_number(parsed_number):
+                return phonenumbers.format_number(
+                    parsed_number, phonenumbers.PhoneNumberFormat.E164
+                )
+    except phonenumbers.NumberParseException:
+        pass
+    return None
 
 
 def blake2b160_hex(text: str) -> str:
